+++ Cargo.toml
@@ -63,6 +63,7 @@ near-jsonrpc-primitives = "0.17.0"
 
 interactive-clap = "0.2.4"
 interactive-clap-derive = "0.2.4"
+keyring = "2.0.5"
 
 [target.'cfg(target_os = "macos")'.dependencies]
 security-framework = "2.7.0"
